+++ internal/convert/testdata/clu2adv/autoscaling.in.tf
