+++ internal/convert/testdata/clu2adv/autoscaling_missing_attribute.in.tf
