+++ internal/convert/testdata/clu2adv/configuration_file_error.in.tf
