+++ internal/convert/testdata/clu2adv/free_cluster_missing_attribute.in.tf
