+++ internal/convert/testdata/clu2adv/free_cluster_with_count.in.tf
